refactor: Line의 구간 처리와 관련된 로직들을 Sections로 이동
