+++ src/main/java/nextstep/subway/line/application/LineService.java
@@ -70,6 +70,6 @@ public class LineService {
     public void removeLineStation(Long lineId, Long stationId) {
         Line line = findLineById(lineId);
         Station station = stationService.findStationById(stationId);
-        line.removeStation(station);
+        line.sections.removeStation(line, station);
     }
 }
