+++ src/main/java/nextstep/subway/line/domain/Line.java
@@ -1,9 +1,6 @@
 package nextstep.subway.line.domain;
 
-import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.List;
-import java.util.Optional;
 import javax.persistence.Column;
 import javax.persistence.Embedded;
 import javax.persistence.Entity;
@@ -25,7 +22,7 @@ public class Line extends BaseEntity {
     private String color;
 
     @Embedded
-    private Sections sections = new Sections();
+    public Sections sections = new Sections();
 
     public Line() {
     }
@@ -58,114 +55,15 @@ public class Line extends BaseEntity {
         return color;
     }
 
-    public List<Section> getSections() {
-        return sections.getSections();
+    public Sections getSections() {
+        return sections;
     }
 
     public void addSection(final Section section) {
-        List<Station> stations = getStations();
-        boolean isUpStationExisted = isStationExists(section.getUpStation());
-        boolean isDownStationExisted = isStationExists(section.getDownStation());
-
-        if (isUpStationExisted && isDownStationExisted) {
-            throw new RuntimeException("이미 등록된 구간 입니다.");
-        }
-
-        if (!stations.isEmpty() && isStationNotExists(section.getUpStation()) &&
-            isStationNotExists(section.getDownStation())) {
-            throw new RuntimeException("등록할 수 없는 구간 입니다.");
-        }
-
-        if (stations.isEmpty()) {
-            sections.addSection(section);
-            return;
-        }
-
-        if (isUpStationExisted) {
-            getUpStationMatchSection(section.getUpStation())
-                .ifPresent(it -> it.updateUpStation(section.getDownStation(), section.getDistance()));
-
-            sections.addSection(section);
-        } else if (isDownStationExisted) {
-            getDownStationMathSection(section.getDownStation())
-                .ifPresent(it -> it.updateDownStation(section.getUpStation(), section.getDistance()));
-
-            sections.addSection(section);
-        } else {
-            throw new RuntimeException();
-        }
+        sections.addSection(section);
     }
 
     public List<Station> getStations() {
-        if (getSections().isEmpty()) {
-            return Arrays.asList();
-        }
-
-        List<Station> stations = new ArrayList<>();
-        Station downStation = findUpStation();
-        stations.add(downStation);
-
-        while (downStation != null) {
-            Optional<Section> nextLineStation = getUpStationMatchSection(downStation);
-            if (!nextLineStation.isPresent()) {
-                break;
-            }
-            downStation = nextLineStation.get().getDownStation();
-            stations.add(downStation);
-        }
-
-        return stations;
-    }
-
-    public Station findUpStation() {
-        Station downStation = getSections().get(0).getUpStation();
-        while (downStation != null) {
-            Optional<Section> nextLineStation = getDownStationMathSection(downStation);
-            if (!nextLineStation.isPresent()) {
-                break;
-            }
-            downStation = nextLineStation.get().getUpStation();
-        }
-
-        return downStation;
-    }
-
-    public Optional<Section> getUpStationMatchSection(final Station upStation) {
-        return getSections().stream()
-            .filter(it -> it.getUpStation() == upStation)
-            .findFirst();
-    }
-
-    public Optional<Section> getDownStationMathSection(final Station downStation) {
-        return getSections().stream()
-            .filter(it -> it.getDownStation() == downStation)
-            .findFirst();
-    }
-
-    public boolean isStationNotExists(final Station station) {
-        return getStations().stream().noneMatch(it -> it == station);
-    }
-
-    public boolean isStationExists(final Station station) {
-        return getStations().stream().anyMatch(it -> it == station);
-    }
-
-    public void removeStation(final Station station) {
-        if (getSections().size() <= 1) {
-            throw new RuntimeException();
-        }
-
-        Optional<Section> upLineStation = getUpStationMatchSection(station);
-        Optional<Section> downLineStation = getDownStationMathSection(station);
-
-        if (upLineStation.isPresent() && downLineStation.isPresent()) {
-            Station newUpStation = downLineStation.get().getUpStation();
-            Station newDownStation = upLineStation.get().getDownStation();
-            Distance newDistance = upLineStation.get().getDistance().add(downLineStation.get().getDistance());
-            sections.addSection(Section.of(this, newUpStation, newDownStation, newDistance.getDistance()));
-        }
-
-        upLineStation.ifPresent(it -> getSections().remove(it));
-        downLineStation.ifPresent(it -> getSections().remove(it));
+        return sections.getStations();
     }
 }
